requirements-dev.txt
-r requirements.txt
black==24.3.0
mypy
pylint
pylint-django
reorder-python-imports
safety
